requirements.txt
playwright==1.40.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
jinja2==3.1.2
Pillow==10.1.0